refactor: conditionally render headbar button
